requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==2.9.1
stripe==1.29.1
Werkzeug==0.11.4
wheel==0.24.0
psycopg2
Flask-OAuth
GeoAlchemy2
alembic
plaid-python
flask-socketio
eventlet
git+https://github.com/wayeasycorp/communityapis.git
flask-cors
flask-failsafe
